requirements.txt
boto3==1.34.143
botocore==1.34.143
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==1.26.19
